chore: update tsconfig with path aliases and refactor imports
