+++ src/app/about/page.tsx
@@ -1,4 +1,4 @@
-import AboutPage from '@/app/components/about-page/AboutPage';
+import AboutPage from '@/components/about-page/AboutPage';
 
 export default function Page() {
   return <AboutPage />;
