+++ src/app/not-found.tsx
@@ -1,4 +1,4 @@
-import NotFoundPage from '@/app/components/not-found/NotFound';
+import NotFoundPage from '@/components/not-found/NotFound';
 
 export default function NotFound() {
   return <NotFoundPage />;
