+++ tsconfig.json
@@ -30,9 +30,9 @@
     ],
     "baseUrl": ".",
     "paths": {
-      "@/*": [
-        "src/*"
-      ]
+      "@/*": ["src/*"],
+      "@/app/*": ["src/app/*"],
+      "@/components/*": ["src/app/components/*"]
     }
   },
   "include": [
